refactor:캘린더 코드 병합
