+++ src/pages/CalendarPage.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { useMediaQuery } from 'react-responsive';
 import Calendar from '../components/Calendar';
 
 /*
@@ -21,17 +22,40 @@ function getDiagnoseData (memberId = '') {
   return diagnoseData;
 };
 
-const CalendarPage = () => {
+const CalenderPage = () => {
+  const isMobile = useMediaQuery({ maxWidth: 899 });
+  const isTablet = useMediaQuery({ minWidth: 900, maxWidth: 1423 });
+  const isDesktop = useMediaQuery({ minWidth: 1424 });
+
   const diagnoseData = getDiagnoseData();
-  // 여기서 필요시 '검사자ID'와 '검사결과링크'가 더 필요할 수 있음.
-  // diagnoseData는 함수로 백엔드로 부터 받아와야 할 것임.
 
   return (
-    <div>
+    <div >
+      {isMobile && (
+        <div style={{ backgroundColor: '#2ecc71', color: 'white', padding: '10px' }}>
+          <h2>Calender - Mobile View</h2>
+          <p>This is the content for smaller screens.</p>
+        </div>
+      )}
+      {isTablet && (
+        <div style={{ backgroundColor: '#e67e22', color: 'white', padding: '20px' }}>
+          <h2>Calender - Tablet View</h2>
+          <p>This is the content for medium-sized screens.</p>
+        </div>
+      )}
+      {isDesktop && (
+        <div style={{ backgroundColor: '#3498db', color: 'white', padding: '30px' }}>
+          <h2>Calender - Desktop View</h2>
+          <p>This is the content for larger screens.</p>
+        </div>
+      )}
+    </div>,
+        <div>
       <h1>Calendar Page</h1>
       <Calendar diagnoseData={diagnoseData} />
     </div>
-  );
-};
 
-export default CalendarPage;
+  );
+}
+
+export default Calender;
